requirements.txt
appdirs==1.4.4
argcomplete==3.6.3
importlib_resources==6.5.2
pillow==12.1.0
pycups==2.0.4
pyserial==3.5
python-barcode==0.16.1
python-escpos==3.1
pyusb==1.3.1
PyYAML==6.0.3
qrcode==8.2
setuptools==80.9.0
six==1.17.0
wheel==0.45.1
